testes para corrida calculada por distancia

+++ DesignPatterns/src/GOF/Creational/Factory_Method/Location.cs
@@ -5,7 +5,7 @@
 		public Coordenada Coordenada { get; private set; }
 		public readonly DateTime Date;
 
-		public Location(decimal latitude, decimal longitude, DateTime date)
+		public Location(double latitude, double longitude, DateTime date)
 		{
 			Date = date;
 			Coordenada = new Coordenada(latitude, longitude);
